+++ CHANGELOG.md
@@ -4,9 +4,13 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 1.0.4 (TBD)
+## [1.1.0](https://github.com/ngageoint/geopackage-tiff-ios/releases/tag/1.1.0) (11-21-2017)
 
-* TBD
+* TIFF Field Type sample format utilities
+* Rasters initializer support for multiple samples per pixel
+* Handle missing samples per pixel with default value of 1
+* Public access to tiff tags
+* String Entry Value getter and setter
 
 ## [1.0.3](https://github.com/ngageoint/geopackage-tiff-ios/releases/tag/1.0.3) (07-10-2017)